post.jsを作成＋Turbolinksの停止
