+++ app/views/posts/show.html.haml
@@ -20,3 +20,6 @@
         .comment__right
           %p.comment__account= comment.user.account
           %p.comment__content= comment.content
+
+
+= javascript_pack_tag 'post'
